Match existing behavior for json file location and name, move logic out of utils into task
